+++ spec/spec_helper.rb
@@ -1,0 +1,5 @@
+require 'minitest/autorun'
+require 'minitest/focus'
+require 'pry'
+
+require_relative '../lib/retriable'